+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Main
 
+* Add support for JDK 15
+
 ## v106
 
 * JDBC_DATABASE_URL query parameters are now alphabetically ordered.