+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Main
 
+## v107
+
 * Add support for JDK 15
 
 ## v106